+++ src/views/guestViews/guestLayout/Header.tsx
@@ -16,7 +16,7 @@ import LanguageDropdown from 'components/common/LanguageDropdown';
 import MoreFilters from '../searchPage/moreFilters';
 import { HeaderDropdownStyledBox, MenuContainer } from './GuestLayout.styled';
 import MenuItem from '@mui/material/MenuItem';
-import { useAuthContext } from '../../../../context/AuthContext';
+import { useAuthContext } from '../../../contexts/AuthContext';
 import { MultipleOptionString } from 'views/protectedModelViews/verification/stepOne/VerificationStepOne';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
