+++ src/views/protectedDashboardViews/dashboardNavItem/HeaderAuthComponent.tsx
@@ -18,9 +18,9 @@ import { Root } from 'services/notification/type';
 import UINewTypography from 'components/UIComponents/UINewTypography';
 import { MODEL_ACTIVE_STEP } from 'constants/workerVerification';
 // import { useCallFeatureContext } from '../../../../context/CallFeatureContext';
-import { useZegoCallFeatureContext } from '../../../../context/ZegoCallContext';
+import { useZegoCallFeatureContext } from '../../../contexts/ZegoCallContext';
 import MyProfileChangePassword from 'views/protectedViews/changePassword';
-import { useAuthContext } from '../../../../context/AuthContext';
+import { useAuthContext } from '../../../contexts/AuthContext';
 import { PAYOUT_ACTION } from 'constants/payoutsConstants';
 import React from 'react';
 import useMediaQuery from '@mui/material/useMediaQuery';
