+++ src/views/protectedModelViews/ModelLastAvtive/index.tsx
@@ -7,7 +7,7 @@ import { ModelDetailsService } from 'services/modelDetails/modelDetails.services
 import { getUserDataClient } from 'utils/getSessionData';
 import { ErrorMessage } from 'constants/common.constants';
 import { User } from 'app/(guest)/layout';
-import { useAuthContext } from '../../../../context/AuthContext';
+import { useAuthContext } from '../../../contexts/AuthContext';
 
 const ModelLastActive = () => {
   const { session } = useAuthContext();
