+++ src/views/protectedViews/bilingDetails/index.tsx
@@ -21,7 +21,7 @@ import {
 import WorkerCardMobile from 'views/guestViews/commonComponents/mobileWorkerCard';
 import { FormattedMessage } from 'react-intl';
 // import { useCallFeatureContext } from '../../../../context/CallFeatureContext';
-import { useZegoCallFeatureContext } from '../../../../context/ZegoCallFeatureContext';
+import { useZegoCallFeatureContext } from '../../../contexts/ZegoCallFeatureContext';
 import moment from 'moment';
 import { useRouter } from 'next/navigation';
 import { ViewDetailsRes } from 'services/guestBilling/types';
